Runnable, but buggy.
Had a lot of trouble updating the dependencies for Flutter Sound.

Not sure what worked, but
- Updated Cocoa Pods, had to install rvm and then Ruby 2.6.6
- Updated Target in iOS runner to 10 as was seeing this - https://github.com/tanersener/flutter-ffmpeg/issues/43
- Tried flutter clean, pod repo update,pod install --repo-update, pod update tau_sound_core

+++ lib/bloc/audio/audio_events.dart
@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:notable/model/audio_note.dart';
 
 abstract class AudioNoteEvent extends Equatable {
@@ -53,7 +52,7 @@ class StopAudioPlaybackRequest extends AudioNoteEvent { }
 
 class AudioRecordingProgressChanged extends AudioNoteEvent {
   final bool isRecording;
-  final double progress;
+  final int progress;
 
   const AudioRecordingProgressChanged(this.isRecording, this.progress);
 
@@ -79,7 +78,7 @@ class AudioRecordingLevelChanged extends AudioNoteEvent {
 
 class AudioPlaybackProgressChanged extends AudioNoteEvent {
   final bool isPlaying;
-  final double progress;
+  final int progress;
 
   const AudioPlaybackProgressChanged(this.isPlaying, this.progress);
